Making it possible to include un-designed pose scores with MergeBB
If you have multiple conformations for the WT (to be fair!) then
you'll need to be able to score those conformations in the fitness
function. This is possible in MSD with the POSE_ENERGIES_VECTOR
command.

What I'm adding here is a way to say in the yaml state version
that there is an additional list file and then to have that file
all of the pdbs inside it get symlinked into the run directory
by the MSDInterfaceDesign class