Add support for POSE_ENERGY_VECTOR structures
For the MergeBB setups, there is good reason to include
the conformations + energies of structures that will not
be designed (e.g. WT monomers) in order to perform accurate
and fair dGbind calculations (i.e. the backbone conformation
available for the design should also be available for the
WT monomer). To support this, the state version yaml file
now lets additional .list files for "pev" entries, and will
symlink all of the .pdb files listed therein.

Also, it looks like another batch of nodes has been added to
dogwood (hurray!) so I'm adjusting the server_identification
script